fix bunch of bugs and inefficencies in vao + gltf

+++ test/assets/shaders/anim.vert
@@ -3,7 +3,7 @@
 layout (location=0) in vec3 aPos;
 layout (location=1) in vec3 aNormal;
 layout (location=2) in vec2 aTexCoords;
-layout (location=3) in vec4 aJoints;
+layout (location=3) in uvec4 aJoints;
 layout (location=4) in vec4 aWeights;
 
 out vec2 texCoords;
